Move column removing logic to TableUtils#removeColumns.
